# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GigWork — Gig-based local work marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone number (mock OTP for web)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Phone already registered
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────────
...
  /v1/jobs/{jobId}/apply:
    post:
      operationId: applyToJob
      tags: [applications]
      summary: Apply to a job (seeker only)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Already applied or job not active
        "403":
          description: Seeker role required
  /v1/jobs/{jobId}/applications:
    get:
      operationId: listJobApplications
      tags: [applications]
      summary: List applicants for a job (provider only, own job)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Applicants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /v1/applications/mine:
    get:
      operationId: listMyApplications
      tags: [applications]
      summary: List current seeker's applications
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, accepted, rejected, withdrawn, completed, null]
      responses:
        "200":
          description: My applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /v1/applications/{applicationId}/accept:
    put:
      operationId: acceptApplication
      tags: [applications]
      summary: Accept an application (provider only)
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /v1/applications/{applicationId}/reject:
    put:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject an application (provider only)
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /v1/applications/{applicationId}/withdraw:
    put:
      operationId: withdrawApplication
      tags: [applications]
      summary: Withdraw own application (seeker only)
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── REVIEWS ──────────────────────────────────────────────────────────────────
...
  /v1/chat/rooms:
    get:
      operationId: listChatRooms
      tags: [chat]
      summary: List chat rooms for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chat rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatRoom"
  /v1/chat/rooms/{roomId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: List messages in a chat room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a chat room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ─────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [phone, name, password, role]
      properties:
        phone:
          type: string
        name:
          type: string
        password:
          type: string
        city:
          type: string
        role:
          type: string
          enum: [seeker, provider]
    AuthResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
        is_new_user:
          type: boolean
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── USER SCHEMAS ─────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, phone, current_role, created_at]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: ["string", "null"]
        profile_photo_url:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        current_role:
          type: string
          enum: [seeker, provider]
        is_phone_verified:
          type: boolean
        is_id_verified:
          type: boolean
        trust_score:
          type: number
        is_banned:
          type: boolean
        seeker_profile:
          $ref: "#/components/schemas/SeekerProfile"
        provider_profile:
          $ref: "#/components/schemas/ProviderProfile"
        created_at:
          type: string
    SeekerProfile:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        availability:
          type: string
          enum: [immediate, within_week, flexible]
        preferred_work_type:
          type: string
          enum: [hourly, daily, monthly, any]
        salary_expectation:
          type: ["integer", "null"]
        years_experience:
          type: integer
        total_jobs_completed:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        attendance_score:
          type: number
        skills:
          type: array
          items:
            type: string
    ProviderProfile:
      type: object
      properties:
        business_name:
          type: ["string", "null"]
        business_type:
          type: ["string", "null"]
        total_jobs_posted:
          type: integer
        total_hires:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        avg_response_hours:
          type: ["number", "null"]
        hiring_success_rate:
          type: number
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        name_hi:
          type: ["string", "null"]
        icon_name:
          type: ["string", "null"]
        sort_order:
          type: integer

    # ── JOB SCHEMAS ──────────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, title, work_type, city, status, created_at]
      properties:
        id:
          type: string
        provider_id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        work_type:
          type: string
          enum: [hourly, daily, monthly]
        # hourly
        hourly_rate:
          type: ["integer", "null"]
        total_hours:
          type: ["integer", "null"]
        shift_start_time:
          type: ["string", "null"]
        shift_end_time:
          type: ["string", "null"]
        same_day_payment:
          type: ["boolean", "null"]
        urgent_hiring:
          type: ["boolean", "null"]
        # daily
        daily_wage:
          type: ["integer", "null"]
        num_working_days:
          type: ["integer", "null"]
        food_included:
          type: ["boolean", "null"]
        accommodation_included:
          type: ["boolean", "null"]
        overtime_available:
          type: ["boolean", "null"]
        # monthly
        monthly_salary:
          type: ["integer", "null"]
        joining_date:
          type: ["string", "null"]
        experience_required:
          type: ["integer", "null"]
        working_days_per_week:
          type: ["integer", "null"]
        pf_esi_included:
          type: ["boolean", "null"]
        salary_negotiable:
          type: ["boolean", "null"]
        # location
        city:
          type: string
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        # status
        status:
          type: string
          enum: [draft, active, paused, filled, expired, removed]
        is_urgent:
          type: boolean
        is_verified:
          type: boolean
        views_count:
          type: integer
        applications_count:
          type: integer
        skills_required:
          type: array
          items:
            type: string
        provider:
          $ref: "#/components/schemas/UserProfile"
        expires_at:
          type: ["string", "null"]
        created_at:
          type: string
        has_applied:
          type: ["boolean", "null"]
...
    Application:
      type: object
      required: [id, job_id, seeker_id, status, applied_at]
      properties:
        id:
          type: string
        job_id:
          type: string
        seeker_id:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn, completed]
        employer_note:
          type: ["string", "null"]
        applied_at:
          type: string
        responded_at:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        job:
          $ref: "#/components/schemas/Job"
        seeker:
          $ref: "#/components/schemas/UserProfile"
        chat_room_id:
          type: ["string", "null"]
    RejectInput:
      type: object
      properties:
        employer_note:
          type: string

    # ── REVIEW SCHEMAS ────────────────────────────────────────────────────────────
...
    ChatRoom:
      type: object
      required: [id, seeker_id, provider_id, created_at]
      properties:
        id:
          type: string
        application_id:
          type: ["string", "null"]
        seeker_id:
          type: string
        provider_id:
          type: string
        last_message_at:
          type: ["string", "null"]
        last_message:
          type: ["string", "null"]
        unread_count:
          type: integer
        other_user:
          $ref: "#/components/schemas/UserProfile"
        job:
          $ref: "#/components/schemas/Job"
        created_at:
          type: string
    ChatMessage:
      type: object
      required: [id, room_id, sender_id, content, created_at]
      properties:
        id:
          type: string
        room_id:
          type: string
        sender_id:
          type: string
        message_type:
          type: string
          enum: [text, system]
        content:
          type: string
        is_read:
          type: boolean
        sender:
          $ref: "#/components/schemas/UserProfile"
        created_at:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── NOTIFICATION SCHEMAS ──────────────────────────────────────────────────────
...